refactor: 파일 구조 정리 및 파일 이동 (#32) (#33)
- 사용하지 않는 파일 및 불필요한 파일 삭제
- 파일 구조 변경에 따른 관련 파일의 경로 업데이트
