+++ app/interfaces/controller/user_controller.go
@@ -1,3 +1,3 @@
-package interfaces
+package controller
 
 // FIXME: 사용자 관련 HTTP 요청을 처리하는 컨트롤러 구현
