+++ app/interfaces/controller/video_job_controller.go
@@ -1,3 +1,3 @@
-package interfaces
+package controller
 
 // FIXME: 동영상 작업 요청을 처리하는 컨트롤러 구현
