+++ app/interfaces/repository/user_repository.go
@@ -1,3 +1,3 @@
-package interfaces
+package repository
 
 // FIXME: 사용자 데이터베이스 액세스 레이어 구현
