+++ app/interfaces/repository/user_video_permission_repository.go
@@ -1,3 +1,3 @@
-package interfaces
+package repository
 
 // FIXME: 사용자-동영상 권한 데이터베이스 액세스 레이어 구현
