+++ app/interfaces/repository/video_job_repository.go
@@ -1,3 +1,3 @@
-package interfaces
+package repository
 
 // FIXME: 동영상 작업 데이터베이스 액세스 레이어 구현
