--- app/interfaces/video_controller.go
@@ -1,3 +1,0 @@
-package interfaces
-
-// FIXME: 동영상 관련 HTTP 요청을 처리하는 컨트롤러 구현
