--- app/interfaces/video_repository.go
@@ -1,3 +1,0 @@
-package interfaces
-
-// FIXME: 동영상 데이터베이스 액세스 레이어 구현
