--- app/usecases/video_repository.go
@@ -1,3 +1,0 @@
-package usecases
-
-// FIXME: 동영상 데이터 액세스 로직 처리 구현
